fix: solve bug when running user endpoints
